excercise 4.21

+++ controllers/blogs.js
@@ -1,6 +1,15 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+/* const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if(authorization && authorization.startsWith('Bearer ')){
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+} */
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -14,6 +23,15 @@ blogRouter.get('/:id', async (request, response) => {
 
 blogRouter.post('/', async (request, response) => {
   const body = request.body
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  console.log('token:', decodedToken)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
   if (body.title === undefined) {
     return response.status(400).json({
       error: 'Error: the title of the blog is missing',
@@ -27,9 +45,6 @@ blogRouter.post('/', async (request, response) => {
 
   body.likes = body.likes === undefined ? 0 : body.likes
 
-  const userList = await User.find({})
-  const user = userList[0]
-
   body.user = user.id
 
   const blog = new Blog(body)
@@ -37,14 +52,19 @@ blogRouter.post('/', async (request, response) => {
 
   response.status(201).json(result)
 
-  console.log('result:', result)
   user.blogs = user.blogs.concat(result.id)
   await user.save()
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === decodedToken.id) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'token invalid' })
+  }
 })
 
 blogRouter.put('/:id', async (request, response, next) => {
